+++ crates/core/tests/java/catch_on_new_line/try_catch_finally_on.out.java
@@ -1,0 +1,15 @@
+class TryCatchFinally {
+    void m() {
+        try {
+            risky();
+        }
+        catch (IOException e) {
+            log(e);
+        }
+        catch (Exception e) {
+            log2(e);
+        } finally {
+            cleanup();
+        }
+    }
+}