+++ crates/core/tests/java/else_on_new_line/else_chain_on.out.java
@@ -1,0 +1,13 @@
+class ElseChain {
+    void m(int a, int b, int c) {
+        if (a) {
+            one();
+        }
+        else if (b) {
+            two();
+        }
+        else {
+            three();
+        }
+    }
+}