+++ crates/core/tests/java/keep_control_statement_in_one_line/own_line_bodies.java
@@ -1,0 +1,13 @@
+class OwnLineBodies {
+    void m(int x, int n, boolean go) {
+        if (x)
+            foo();
+        while (go)
+            step();
+        for (int i = 0; i < n; i++)
+            use(i);
+        do
+            tick();
+         while (go);
+    }
+}