+++ crates/core/tests/java/while_on_new_line/do_while.java
@@ -1,0 +1,10 @@
+class DoWhile {
+    void m(boolean go) {
+        do {
+            tick();
+        } while (go);
+        do
+            tick();
+         while (go);
+    }
+}